refactor: ubuntu_provision_test for UDI
