+++ packages/ubuntu_provision_test/lib/ubuntu_provision_test.dart
@@ -1,5 +1,7 @@
 library ubuntu_provision_test;
 
+export 'src/bootstrap_tester.dart';
 export 'src/expect.dart';
 export 'src/fake_init.dart';
-export 'src/page_tester.dart';
+export 'src/provision_tester.dart';
+export 'src/wizard_tester.dart';
